+++ src/app/collective-rewards/components/CycleMetrics/CycleEndingOn.tsx
@@ -1,0 +1,10 @@
+import { LoadingSpinner } from '@/components/LoadingSpinner'
+import { Metric } from '@/components/Metric'
+import { DateTime } from 'luxon'
+
+export const CycleEndingOn = ({ cycleNext, isLoading }: { cycleNext: DateTime; isLoading: boolean }) => {
+  if (isLoading) {
+    return <LoadingSpinner />
+  }
+  return <Metric title="Cycle ending on"> {cycleNext.toFormat('EEE, dd MMM')}</Metric>
+}
