+++ src/app/collective-rewards/components/CycleMetrics/index.ts
@@ -1,0 +1,1 @@
+export * from './CycleMetrics'
