+++ src/app/collective-rewards/components/EstimatedRewards/index.ts
@@ -1,0 +1,1 @@
+export * from './EstimatedRewards'
