+++ src/app/collective-rewards/components/TotalBacking/index.ts
@@ -1,0 +1,1 @@
+export * from './TotalBacking'
